% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwflandscapetheme.R
\name{custom_fill}
\alias{custom_fill}
\title{Custom fill scale for scatterplots (if needed for filled points or areas)}
\usage{
custom_fill()
}
\value{
A ggplot2 fill scale.
}
\description{
Custom fill scale for scatterplots (if needed for filled points or areas)
}
