% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwflandscapetheme.R
\name{custom_points}
\alias{custom_points}
\title{Custom points layer for square markers}
\usage{
custom_points(size = 3)
}
\arguments{
\item{size}{Numeric value for point size (default is 3).}
}
\value{
A ggplot2 layer for square-shaped points.
}
\description{
This function creates a ggplot2 layer with square-shaped points.
}
