% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwflandscapetheme.R
\name{horizontal_bar_theme}
\alias{horizontal_bar_theme}
\title{Custom Theme for Horizontal Bar Graphs}
\usage{
horizontal_bar_theme()
}
\value{
A ggplot2 theme object optimized for horizontal bar charts.
}
\description{
Custom Theme for Horizontal Bar Graphs
}
